requirements.txt
Django==4.2.15
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.124.0
python-dotenv==1.0.0